+++ 3-linked-list/list.h
@@ -6,14 +6,12 @@
 // TODO: Complementary work needed: Functions that doesn't modify the
 // list should be declared const.
 //
-// TODO: Complementary work needed: The list should contain ints, and not double.
-//
 // TODO: Complementary work needed: Missing print member function.
 
 struct Element {
-  double value;
+  int value;
   Element *next;
-  Element(double value) : value(value), next(nullptr) {}
+  Element(int value) : value(value), next(nullptr) {}
 };
 
 class List {
@@ -22,17 +20,17 @@ private:
 
 public:
   List() : first(nullptr){};
-  List(std::initializer_list<double> List);
+  List(std::initializer_list<int> List);
   List(const List &other);
   List(List &&other);
   ~List();
   List &operator=(List const &other);
   List &operator=(List &&other);
-  void insert(double value);
-  double find_value_with_rank(int rank);
-  int find_rank_with_value(double value);
+  void insert(int value);
+  int find_value_with_rank(int rank);
+  int find_rank_with_value(int value);
   void delete_element_with_rank(int rank);
-  void delete_element_with_value(double value);
+  void delete_element_with_value(int value);
   int size();
   bool is_empty();
   std::string to_string() const;
